ci: add test and lint workflows
This commit adds continuous integration capability to this repo, so
tests can be run automatically when added.

`golangci-lint` seems to be the standard linter for `golang`, and this
linter have already been used in several other repos so that is why it
was chosen.

The sub-linters that are enabled are the by default enabled ones, as of
version v1.53.3:

* `errcheck`
* `gosimple`
* `govet`
* `ineffassign`
* `staticcheck`
* `typecheck`
* `unused`

of the default disabled ones the following has been enabled, as of
version v1.53.3:

* `gofmt`, to ensure all code has been formatted properly

An [issue](https://github.com/nlnwa/go_container/issues/3) has been
introduced to deal with the remaining sub-linters.