refactor: use minimal docker image as run-stage
Using a minimal image as a base for the `run`-stage in the docker file
ensures that there would be minimal dependencies between the `build` and
the `run`.

By using `distroless` images from
https://github.com/GoogleContainerTools/distroless we run the code in a
very minimal image.

`base` seems more restrictive than `static`, so setting it to `base`
seems like a good default.